refactor!: Re-impl all macro to derive
The former impl use attribute macro, which needs
to keep the original input. Since we don't need to
modify the input, it is better to use derive
instead.

In addition, all impl are split into separate
modules. Only the exported macro and doc are in
the main lib.rs.

bump to v0.0.3

BREAKING CHANGE: The macro is now derive macro